refactor: port PlatformCard and PlatformExplorer to Svelte 5
Replace the SolidJS PlatformCard and PlatformExplorer islands with
Svelte 5 rune-based equivalents, matching the existing ThemeToggle.svelte
house style. PlatformCard is ported first since PlatformExplorer renders
it as a child.

- createSignal -> $state, createMemo -> $derived/$derived.by
- <For>/<Show>/<Switch> -> {#each}/{#if}, each loop keyed by a stable id
- value+onInput/onChange -> bind:value (identical event semantics)
- classList -> inline ternary class expression, preserving the exact
  literal UnoCSS class strings so extraction still finds them
- mergeProps default handling -> destructured $props() defaults

Prop contracts (platforms/currentGrade/searchLabel) are unchanged; the
three Astro pages that mount PlatformExplorer only get their import
extension updated, client:idle directives untouched. Rendered markup,
Danish UI strings, and behaviour are otherwise identical.
