+++ src/pages/fag/[subject].astro
@@ -1,6 +1,6 @@
 ---
 import { getCollection } from "astro:content";
-import PlatformExplorer from "../../components/PlatformExplorer.tsx";
+import PlatformExplorer from "../../components/PlatformExplorer.svelte";
 import Layout from "../../layouts/Layout.astro";
 import { getPlatformsForSubject } from "../../lib/platforms";
 import { subjectMeta } from "../../lib/subjects";
