+++ src/pages/grade/[id].astro
@@ -1,6 +1,6 @@
 ---
 import GradeRail from "../../components/GradeRail.astro";
-import PlatformExplorer from "../../components/PlatformExplorer.tsx";
+import PlatformExplorer from "../../components/PlatformExplorer.svelte";
 import Layout from "../../layouts/Layout.astro";
 import { GRADES, gradeLabel, type PhaseId, phaseOf } from "../../lib/grades";
 import { getPlatformsForGrade } from "../../lib/platforms";
