+++ src/pages/other.astro
@@ -1,5 +1,5 @@
 ---
-import PlatformExplorer from "../components/PlatformExplorer.tsx";
+import PlatformExplorer from "../components/PlatformExplorer.svelte";
 import Layout from "../layouts/Layout.astro";
 import { getOtherPlatforms } from "../lib/platforms";
